move files to `app`
