+++ src/app/personalization/api/cart/add-item/route.ts
@@ -1,4 +1,4 @@
-import { ProductDbModel, UserCartItemDbModel } from '../../../../../server/personalization/database';
+import { ProductDbModel, UserCartItemDbModel } from '../../database';
 import { Op } from 'sequelize';
 import { NextRequest, NextResponse } from 'next/server';
 import { getAndValidateFingerprintResult, Severity } from '../../../../../server/checks';
