+++ src/app/personalization/components/productCard.module.scss
