+++ src/app/personalization/components/searchComponents.module.scss
