+++ src/app/personalization/hooks/use-products.ts
