--- src/client/components/snackbar-action.tsx
@@ -1,12 +1,0 @@
-import { SnackbarKey, useSnackbar } from 'notistack';
-import Button from '@mui/material/Button';
-
-export function SnackbarAction({ snackbarId }: { snackbarId: SnackbarKey }) {
-  const { closeSnackbar } = useSnackbar();
-
-  return (
-    <Button onClick={() => closeSnackbar(snackbarId)} color='inherit'>
-      Close
-    </Button>
-  );
-}
